# api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backup Manager /api/v1",
    "version": "1.0.0",
    "summary": "The authoritative definition of the versioned /api/v1 boundary.",
    "description": "Issue #166 (B6.2). This document, not the Go handlers and not the TypeScript client, is the definition of /api/v1. Both sides are generated from it (scripts/api/generate.sh) and both are held to it by a drift check (scripts/api/check-contract-drift.sh plus the conformance tests named in docs/api/contract.md). Compatibility and deprecation rules are in docs/api/contract.md."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Every path in this document is relative to the versioned base path. /health/live and /health/ready are deliberately NOT in this contract: they sit outside /api/v1, carry no authentication and exist for an orchestrator, not for a client of this API."
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bm_session",
        "description": "The authenticated session. For the generic runtime profile this is apps/common/auth/local's own HTTP-only cookie; a runtime profile with a native gateway (native_auth) establishes the same session from provider identity headers accepted only from a trusted peer. Which mechanism established it is a profile concern; that a caller of a secured operation has one is not."
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "The double-submit CSRF token (apps/common/csrf). Required on every state-changing operation, and on the two read-only-in-effect operations that still open an outbound connection to a caller-supplied host."
      },
      "bootstrapToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Bootstrap-Token",
        "description": "The single-use enrollment secret printed to the runtime's own log. Only /auth/enroll accepts it."
      }
    },
...
    "schemas": {
...
      "BackupSet": {
        "type": "object",
        "description": "A persisted backup set as the API reports it.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "source_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "type": "string"
          },
          "remote_path": {
            "type": "string"
          },
          "local_path": {
            "type": "string"
          },
          "include": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "completion_strategy": {
            "type": "string",
            "enum": [
              "rename",
              "marker",
              "stable"
            ]
          },
          "validator_id": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "source_name",
          "name",
          "host",
          "port",
          "user",
          "remote_path",
          "local_path",
          "include",
          "completion_strategy",
          "validator_id",
          "disabled"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The nested error body every operation outside /auth returns. code is stable and machine-readable; message is human-readable and MAY change without notice.",
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "The error envelope every operation outside /auth returns. The correlation id travels in the X-Correlation-Id response header.",
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Artifact": {
        "type": "object",
        "description": "One backup this deployment holds, as the API reports it. Nothing here names an implementation: the fields describe a backup's identity, freshness and trustworthiness, never how it is stored.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "The three-part \"source/set/name\" identity."
          },
          "backup_set_id": {
            "type": "string",
            "description": "The two-part \"source/set\" identity of the backup set this belongs to."
          },
          "source_name": {
            "type": "string"
          },
          "set_name": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "description": "The artifact's own filename within its backup set. Always a basename, never a path."
          },
          "remote_path": {
            "type": "string"
          },
          "local_path": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "description": "The lifecycle state this backup is currently recorded in, e.g. \"DISCOVERED\", \"COMMITTED\", \"COMPLETE\", \"QUARANTINED\"."
          },
          "discovered_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "size_bytes": {
            "type": "integer",
            "description": "What the source reported at discovery, or 0 when it reported no size at all.",
            "format": "int64"
          },
          "checksum": {
            "type": "string",
            "description": "The hash recorded for the durable local copy at verification. Empty when nothing has been hashed.",
            "x-go-omitempty": true
          },
          "checksum_algorithm": {
            "type": "string",
            "x-go-omitempty": true
          },
          "validation": {
            "type": "string",
            "description": "Whether this backup's content has been validated. Spelled out rather than left as a nullable boolean, so a caller never has to decide what an absent value means.",
            "enum": [
              "passed",
              "failed",
              "pending"
            ]
          },
          "validation_detail": {
            "type": "string",
            "x-go-omitempty": true
          },
          "remote_source_removed_at": {
            "type": "string",
            "description": "When the source object was released. Absent while it is still there. Releasing a source is a lifecycle fact, never something a client asks for.",
            "format": "date-time",
            "x-go-omitempty": true
          },
          "quarantined": {
            "type": "boolean",
            "description": "True when this backup is not trusted and is being held for a human."
          },
          "quarantine_irrecoverable": {
            "type": "boolean",
            "description": "True when no source remains anywhere to re-ingest from, so retrying cannot help."
          },
          "quarantine_reason": {
            "type": "string",
            "x-go-omitempty": true
          },
          "retention_tier": {
            "type": "string",
            "description": "The retention tier that most recently selected this backup to be kept.",
            "x-go-omitempty": true
          }
        },
        "required": [
          "id",
          "backup_set_id",
          "source_name",
          "set_name",
          "name",
          "remote_path",
          "local_path",
          "state",
          "discovered_at",
          "updated_at",
          "size_bytes",
          "validation",
          "quarantined",
          "quarantine_irrecoverable"
        ]
      },
      "ListArtifactsResponse": {
        "type": "object",
        "description": "GET /backups and GET /quarantine. An object with one array field, never a bare top-level array.",
        "additionalProperties": false,
        "properties": {
          "artifacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Artifact"
            }
          }
        },
        "required": [
          "artifacts"
        ]
      },
      "ArtifactCheckResponse": {
        "type": "object",
        "description": "The verdict of re-checking one quarantined backup. Nothing is written and the backup is not moved: a pass is evidence for an operator deciding whether to retry, not a rehabilitation.",
        "additionalProperties": false,
        "properties": {
          "checked": {
            "type": "boolean",
            "description": "False when there was nothing to examine at all, which is still not a pass."
          },
          "passed": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "description": "What was checked and what it found, already a sentence.",
            "x-go-omitempty": true
          }
        },
        "required": [
          "checked",
          "passed"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/backup-sets/{id}": {
      "get": {
        "operationId": "getBackupSet",
        "summary": "Read one backup set. The id is a two-part source/name identity, so it spans path segments.",
        "tags": [
          "backup-sets"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The two-part identity, e.g. \"nas-a/photos\". Registered in the router as a catch-all for exactly that reason."
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupSet"
                }
              }
            }
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "404": {
            "description": "Refused: BACKUP_SET_NOT_FOUND.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "BACKUP_SET_NOT_FOUND"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": false,
        "x-authenticated": true
      }
    },
...
    "/backups": {
      "get": {
        "operationId": "listArtifacts",
        "summary": "Every backup this deployment holds, optionally narrowed to one backup set.",
        "tags": [
          "backups"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "setId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "A two-part \"source/set\" identity. An id that names no configured backup set selects nothing, which is not an error."
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListArtifactsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": false,
        "x-authenticated": true
      }
    },
    "/backups/{id}": {
      "get": {
        "operationId": "getArtifact",
        "summary": "Read one backup. The id is a three-part source/set/name identity, so it spans path segments.",
        "tags": [
          "backups"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The three-part identity, e.g. \"nas-a/photos/2026-08-30.tar\". Registered in the router as three parameters for exactly that reason."
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artifact"
                }
              }
            }
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "404": {
            "description": "Refused: ARTIFACT_NOT_FOUND.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "ARTIFACT_NOT_FOUND"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": false,
        "x-authenticated": true
      }
    },
...
    "/quarantine": {
      "get": {
        "operationId": "listQuarantine",
        "summary": "Every backup being held for a human, and why.",
        "tags": [
          "quarantine"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListArtifactsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": false,
        "x-authenticated": true
      }
    },
    "/quarantine/{id}/revalidate": {
      "post": {
        "operationId": "revalidateArtifact",
        "summary": "Re-check one quarantined backup and report the verdict. Writes nothing and moves nothing.",
        "tags": [
          "quarantine"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The three-part source/set/name identity."
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactCheckResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "403": {
            "description": "Refused: CSRF_TOKEN_MISSING, CSRF_TOKEN_MISMATCH.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "CSRF_TOKEN_MISSING",
              "CSRF_TOKEN_MISMATCH"
            ]
          },
          "404": {
            "description": "Refused: ARTIFACT_NOT_FOUND.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "ARTIFACT_NOT_FOUND"
            ]
          },
          "409": {
            "description": "Refused: ARTIFACT_NOT_QUARANTINED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "ARTIFACT_NOT_QUARANTINED"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": true,
        "x-authenticated": true
      }
    },
    "/quarantine/{id}/retry": {
      "post": {
        "operationId": "retryArtifactIngestion",
        "summary": "Return one quarantined backup to the pipeline so it is attempted again.",
        "tags": [
          "quarantine"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The three-part source/set/name identity."
          }
        ],
        "responses": {
          "204": {
            "description": "Success. No content."
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "403": {
            "description": "Refused: CSRF_TOKEN_MISSING, CSRF_TOKEN_MISMATCH.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "CSRF_TOKEN_MISSING",
              "CSRF_TOKEN_MISMATCH"
            ]
          },
          "404": {
            "description": "Refused: ARTIFACT_NOT_FOUND.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "ARTIFACT_NOT_FOUND"
            ]
          },
          "409": {
            "description": "Refused: ARTIFACT_NOT_QUARANTINED, ARTIFACT_IRRECOVERABLE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "ARTIFACT_NOT_QUARANTINED",
              "ARTIFACT_IRRECOVERABLE"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": true,
        "x-authenticated": true
      }
    },
...
  },
...
}